Don't fail e2e tests if factory object is null

+++ selenium/che-selenium-test/src/test/java/org/eclipse/che/selenium/factory/DirectUrlFactoryWithRootFolderTest.java
@@ -25,6 +25,8 @@ import org.eclipse.che.selenium.core.factory.TestFactory;
 import org.eclipse.che.selenium.core.factory.TestFactoryInitializer;
 import org.eclipse.che.selenium.pageobject.theia.TheiaIde;
 import org.eclipse.che.selenium.pageobject.theia.TheiaProjectTree;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.testng.annotations.AfterClass;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
@@ -32,6 +34,8 @@ import org.testng.annotations.Test;
 /** @author Musienko Maxim */
 @Test(groups = {GITHUB, OPENSHIFT})
 public class DirectUrlFactoryWithRootFolderTest {
+  private static final Logger LOG =
+      LoggerFactory.getLogger(DirectUrlFactoryWithRootFolderTest.class);
 
   @Inject private TestFactoryInitializer testFactoryInitializer;
   @Inject private TestGitHubRepository testRepo;
@@ -52,6 +56,11 @@ public class DirectUrlFactoryWithRootFolderTest {
 
   @AfterClass
   public void tearDown() throws Exception {
+    if (testFactoryWithRootFolder == null) {
+      LOG.warn("It was impossible to remove factory.");
+      return;
+    }
+
     testFactoryWithRootFolder.delete();
   }
